requirements.txt
aiocache==0.12.3
aiohappyeyeballs==2.4.4
aiohttp==3.11.8
aiosignal==1.3.2
alembic==1.14.0
annotated-types==0.7.0
anthropic==0.45.0
appdirs==1.4.4
APScheduler==3.10.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==5.0.1
attrs==25.1.0
Authlib==1.3.2
av==12.3.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
bitarray==3.0.0
black==24.8.0
blinker==1.9.0
blis==1.2.0
boto3==1.35.53
botocore==1.35.99
build==1.2.2.post1
cachetools==5.5.1
catalogue==2.0.10
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.2
click==8.1.8
cloudpathlib==0.20.0
colbert-ai==0.2.21
colorama==0.4.6
colorclass==2.2.2
coloredlogs==15.0.1
compressed_rtf==1.0.6
confection==0.1.5
convert_numbers==0.4
ctranslate2==4.5.0
cymem==2.0.11
dataclasses-json==0.6.7
datasets==3.2.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.17
dill==0.3.8
Django==5.1.6
django-environ==0.12.0
dnspython==2.7.0
docker==7.1.0
docx2txt==0.8
duckduckgo_search==7.2.1
durationpy==0.9
easygui==0.98.3
ebcdic==1.1.1
ecdsa==0.19.0
einops==0.8.0
email_validator==2.2.0
emoji==2.14.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
eval_type_backport==0.2.2
Events==0.5
executing==2.0.1
extract-msg==0.52.0
fake-useragent==1.5.1
fastapi-cli==0.0.7
faster-whisper==1.0.3
filelock==3.17.0
filetype==1.2.0
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==25.1.24
fonttools==4.55.6
fpdf2==2.8.2
frozenlist==1.5.0
fs==2.4.16
fsspec==2024.9.0
ftfy==6.2.3
gcp-storage-emulator==2024.8.3
git-python==1.0.3
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.6
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-generativeai==0.7.2
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
greenlet==3.1.1
grpcio==1.67.1
grpcio-status==1.62.3
grpcio-tools==1.62.3
h11==0.14.0
h2==4.1.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.4.0
importlib_resources==6.5.2
iniconfig==2.0.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonpath-python==1.0.6
kubernetes==32.0.0
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.31
langchain-text-splitters==0.3.5
langcodes==3.5.0
langdetect==1.0.9
langfuse==2.44.0
langsmith==0.1.147
language_data==1.3.0
lark==1.1.9
ldap3==2.9.1
lxml==5.3.0
Mako==1.3.8
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
moto==5.0.27
mpmath==1.3.0
msal==1.31.1
msoffcrypto-tool==5.4.2
multidict==6.1.0
multiprocess==0.70.16
multitasking==0.0.11
murmurhash==1.0.12
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.3
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
oletools==0.60.2
onnxruntime==1.20.1
open-webui==0.5.7
opencv-python==4.11.0.86
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
opensearch-py==2.7.1
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
passlib==1.7.4
pathspec==0.12.1
pcodedmp==1.2.6
pdf2image==1.17.0
peewee==3.17.8
peewee-migrate==1.12.2
pgvector==0.3.5
pillow==11.1.0
pluggy==1.5.0
portalocker==2.10.1
posthog==3.10.0
preshed==3.0.9
primp==0.11.0
prompt_toolkit==3.0.47
propcache==0.2.1
proto-plus==1.25.0
psutil==6.1.1
psycopg2-binary==2.9.9
py-partiql-parser==0.6.1
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post6
pydantic==2.9.2
pydantic-settings==2.7.1
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.10.1
pymdown-extensions==10.11.2
pymilvus==2.5.0
pymongo==4.10.1
PyMySQL==1.1.1
pypandoc==1.13
pyparsing==3.2.1
pypdf==4.3.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pytesseract==0.3.13
pytest-docker==3.1.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-engineio==4.11.2
python-iso639==2025.1.27
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.18
python-oxmsg==0.0.1
python-pptx==1.0.0
python-socketio==5.11.3
pytube==15.0.0
pytz==2024.1
pywin32==308
pyxlsb==1.0.10
PyYAML==6.0.2
qdrant-client==1.12.2
rank-bm25==0.2.2
RapidFuzz==3.11.0
rapidocr-onnxruntime==1.3.24
red-black-tree-mod==1.20
redis==5.2.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
responses==0.25.6
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
RTFDE==0.1.2
s3transfer==0.10.4
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
sentence-transformers==3.3.1
sentencepiece==0.2.0
setuptools==75.8.0
shapely==2.0.6
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.32
sqlparse==0.5.3
srsly==2.5.1
stack-data==0.6.3
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
thinc==8.3.4
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.1
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
unstructured==0.15.9
unstructured-client==0.28.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.30.6
validators==0.34.0
wasabi==1.1.3
watchfiles==1.0.4
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
wheel==0.45.1
win_unicode_console==0.5
wrapt==1.17.2
wsproto==1.2.0
xlrd==2.0.1
XlsxWriter==3.2.1
xmltodict==0.14.2
xxhash==3.5.0
yarl==1.18.3
yfinance==0.2.41
youtube-transcript-api==0.6.3
zipp==3.21.0
